As a sanity check on ~, I attempted to prove that it's decidable in the sense that for any pair of types you can show that they are either ~ or not ~ but not both. This should be true but the proof didn't go through.
The intuition for why is that TSym had a sort of impredicative nature. When
pattern matching ruled out all the other constructors, TSym will always
still apply, so you can't get a contradiction in the cases like num ~ _ ==>
_ .

I talked to Cyrus about this and we fixed it by changing the definition
slightly. Instead of having a sym rule and one hole rule, we follow the
idea of the gradual typing paper and have two hole rules, which are
symmetric. Then, we can recover symmetry as a metatheorem about the rules,
which we do in the Agda source. This then allows the proof about ~'s
decidability to go through easily.

We also added a more general form of reflexivity instead of the specific
thing at the base type num, since any type is compatible with itself.

All of these will need to change when new types are added to the core
calculus, but that's a good thing. It means you have to find (certail
classes of) bugs in your rules!