fixed the three holes i couldn't figure out in actsense1 after talking to cyrus. they are not counter examples to determinism, nor is the thing i'd written down.
the thing i wrote down isn't because it uses different actions. determinism
says that under the same action the same expression goes to only one
thing. so that was just a mistake on my part.

the way to fix the theorem is to prove that we also have unicity of the
type system--effectively what we're determinism for the action judgements
but for the typing judgements. that does indeed fill the holes.

it creates the new obligation of that proof, though. as a case in that
proof, you need to show that two applications of SVar result in the same
thing. that is not true about the system now, because you can have
duplicates in the context. ∈ picks the first one but it's not strong enough
to prove this. for instance, there's nothing wrong with the context

   Γ := (5, num) :: (5, <||>) :: []

and (5, num) ∈ Γ and (5 , <||>) ∈ Γ, but num is emphatically not <||>.

so there's an upcoming refactor there to use lists that are statically
enforced to be unique up π1 (or any function) so that that case goes through.