rewrote everything that touches contexts to use a new type. contexts are, really, a finite map from vars to types. instead of implementing that as lists of pairs with a bunch of side conditions, i now implement it as a function from variables to possible types.
this tightens up a lot of the code that messes with contexts, as well as
lets context unicity go through which it didn't before (because it was
false). so that's good.

we'll see if this is a good idea forever. it's now very nasty to try to
reason about whether two contexts are equal, because you might have to use
function extensionality. if contexts were also paired up with a "next
unused name" you could get around that by testing from 0 to that and saying
that that's enough to make them equal. you could actually also have a type
of functions that have a baked-in fixed point -- say that f x is None for
all x > some y, in the type. so really finite maps, not infinite maps that
happen to really be finite but don't reflect it in their type.

there's also finite maps in one of the std libs, implemented i think with
AVL tress. i'm reluctant to start mucking with that, though, because it's
often tempting but nasty.