proof of well-modedness for the build jugement triplet complete. note that this means precious little, though. all it tells us is that the judgment has the mode; it does not say anything whatsoever about the correctness of the list that's actually produced. i've written it in a style to exaggerate this, using _ and implicit arguments throughout to avoid needing to ever actually name any of the lists build in the proof term. this means that i can choose whatever lists i want and never need to change the mode argument as long as i don't delete any constructors.
the entire notion of correctness is in the subsequent theorems that say
anything about what happens when you actually run the lists you
produce. indeed as of this commit, several of the lists produced are
patently wrong, but the judgement is still well-moded.

it's also the case that this judgement triplet really has mode (\forall, !
\exists), making it really a function. it's an easy proof and i'll do it
later if i have to, but i hope to avoid it.